Updated controllers /api and routing
