+++ index.js
@@ -1,14 +1,16 @@
 const express = require('express');
 const bcrypt = require('bcrypt');
 const sequelize = require('./config/connection');
-const User = require('./models/User')
+const User = require('./models/User');
 const withAuth = require('./utils/auth');
-const RegisterRoute = require('./controllers/routes/registerRoute');
 const Expense = require('./models/Expense');
 const crypto = require('crypto');
-
 const session = require('express-session');
 const SequelizeStore = require('connect-session-sequelize')(session.Store);
+const path = require('path');
+
+// Use path.join for controllers import
+const controllers = require('./controllers');
 
 const app = express();
 
@@ -28,6 +30,9 @@ const secretKey = generateSecretKey();
 app.use(
   session({
     secret: secretKey,
+    cookie: {
+      maxAge: 24 * 60 * 60 * 1000, // EXPIRES IN 1 DAY
+    },
     resave: false,
     saveUninitialized: true,
     store: new SequelizeStore({
@@ -40,136 +45,22 @@ const hbs = exphbs.create({
   partialsDir: 'views/partials/',
 });
 
+app.use(express.static(__dirname));
+
+app.engine('handlebars', hbs.engine);
+app.set('view engine', 'handlebars');
+
+app.set('views', './views')
+
+app.use('/', controllers);
+
 async function startApp() {
   try {
     console.log('Connection to the database has been established successfully.');
     await sequelize.sync();
     console.log('Database synchronized.');
 
-    app.use(express.static(__dirname));
 
-    app.engine('handlebars', hbs.engine);
-    app.set('view engine', 'handlebars');
-
-    app.set('views', './views')
-
-    app.get('/signup', (req, res) => {
-      res.render('partials/login');
-    });
-
-    app.get('/dashboard', withAuth, async (req, res) => {
-      try {
-        const userData = await User.findByPk(req.session.user_id, {
-          attributes: { exclude: ['password'] },
-          include: [{ model: Expense }],
-        });
-    
-        if (!userData) {
-          res.redirect('/login'); 
-          return;
-        }
-    
-        const user = userData.get({ plain: true });
-    
-        console.log('User Data:', user);
-    
-        res.render('dashboard', {
-          username: user.username,
-          expenses: user.expenses,
-          isAuthenticated: true,
-        });
-      } catch (err) {
-        console.error(err);
-        res.status(500).json(err);
-      }
-    });
-
-    app.get('/login', (req, res) => {
-      if (req.session.logged_in) {
-        res.redirect('/dashboard');
-        return;
-      }
-
-      res.render('/login');
-    });
-
-    app.get('/', (req, res) => {
-      res.redirect('/signup');
-    })
-
-    app.post('/register', async (req, res) => {
-      const { username, password } = req.body;
-
-      try {
-        const newUser = await User.create({
-          username: username,
-          password: password,
-        });
-
-        req.session.user_id = newUser.id;
-        req.session.logged_in = true;
-
-        res.status(201).json(newUser);
-        console.log("New user added successfully!");
-      } catch (error) {
-        console.error(error);
-        res.status(500).json({ error: 'Internal Server Error' });
-      };
-    });
-
-    app.post('/signup', async (req, res) => {
-      const { username, password } = req.body;
-    
-      try {
-        const userData = await User.findOne({
-          where: {
-            username: req.body.username,
-          },
-        });
-    
-        if (!userData) {
-          res.redirect('/signup')
-          return;
-        }
-    
-        const validatePassword = bcrypt.compareSync(password, userData.password);
-    
-        if (!validatePassword) {
-          return;
-        }
-    
-        console.log('Successfully signed in');
-        req.session.user_id = userData.id;
-        req.session.logged_in = true;
-    
-        req.session.save(() => {
-          res.redirect('/dashboard');
-        });
-      } catch (error) {
-        console.error(error);
-        res.status(500).json({ error: 'Internal Server Error' });
-      }
-    });
-
-    app.post('/expenses', async (req, res) => {
-      console.log('Hit /expenses route'); // Add this line
-      try {
-        const { description, spending, date_created, name } = req.body;
-        const user_id = req.session.user_id;
-        const newExpense = await Expense.create({
-          name,
-          description,
-          spending,
-          date_created,
-          user_id,
-      });
-    
-        res.status(201).json(newExpense);
-      } catch (error) {
-        console.error('Error creating expense:', error);
-        res.status(500).json({ error: 'Internal Server Error' });
-      }
-    });
 
     app.listen(PORT, () => {
       console.log(`LISTENING ON PORT ${PORT}`);
